change target column type to bigint

+++ db/migrate/20231204113057_change_column_type.rb
@@ -1,5 +1,5 @@
 class ChangeColumnType < ActiveRecord::Migration[7.0]
   def change
-    change_column :deals, :target, :integer
+    change_column :deals, :target, :bigint
   end
 end
